requirements.txt
Flask==3.0.0
PyYAML==6.0.1
requests==2.31.0
gunicorn==21.2.0
fpdf2==2.7.9
coverage==7.13.4
